// src/store/accounts/actions/types.ts
export enum Constants {
  GET_ACCOUNTS_SUCCESS = 'GET_ACCOUNTS_SUCCESS',
  GET_ACCOUNTS_ERROR = 'GET_ACCOUNTS_ERROR',
  GET_ACCOUNTS_LOADING = 'GET_ACCOUNTS_LOADING',
  CREATE_ACCOUNT_LOADING = 'CREATE_ACCOUNT_LOADING',
  CREATE_ACCOUNT_SUCCESS = 'CREATE_ACCOUNT_SUCCESS',
  CREATE_ACCOUNT_ERROR = 'CREATE_ACCOUNT_ERROR',
  TRANSFER_MONEY_LOADING = 'TRANSFER_MONEY_LOADING',
  TRANSFER_MONEY_SUCCESS = 'TRANSFER_MONEY_SUCCESS',
  TRANSFER_MONEY_ERROR = 'TRANSFER_MONEY_ERROR',
}